EKS Fix (#27)
* Typos in paths

* Testing change in provider config

* EKS split

* Fixing some deps

* Fix in code generation

* Fixing outputs

* Fixing Bitops Execution if nothing is needed

+++ operations/_scripts/generate/generate_bitops_config.sh
@@ -81,6 +81,8 @@ targets="$targets
 targets_attribute="$targets_attribute $targets"
 
 create_bitops_terraform_config aws $AWS_EC2_INSTANCE_CREATE targets
+create_bitops_terraform_config eks $AWS_EKS_CREATE
+
 
 #Will add the user_data file into the EC2 Terraform folder
 if [[ $(alpha_only "$AWS_EC2_INSTANCE_CREATE") == true ]]; then
@@ -93,7 +95,7 @@ fi
 if [[ $(alpha_only "$AWS_EKS_CREATE") == true ]]; then
   if [ -s "$GITHUB_WORKSPACE/$AWS_EKS_INSTANCE_USER_DATA_FILE" ] && [ -f "$GITHUB_WORKSPACE/$AWS_EKS_INSTANCE_USER_DATA_FILE" ]; then
       echo "Moving $AWS_EKS_INSTANCE_USER_DATA_FILE to be used by Terraform during EKS Nodes creation"
-      mv "$GITHUB_WORKSPACE/$AWS_EKS_INSTANCE_USER_DATA_FILE" "$GITHUB_ACTION_PATH/operations/deployment/terraform/aws/aws_eks_incoming_user_data_script.sh"
+      mv "$GITHUB_WORKSPACE/$AWS_EKS_INSTANCE_USER_DATA_FILE" "$GITHUB_ACTION_PATH/operations/deployment/terraform/eks/aws_eks_incoming_user_data_script.sh"
   fi
 fi
 # Below we will be creating the config file, one for the action itself, other to store as an artifact after. 
@@ -103,7 +105,7 @@ mkdir -p "${GITHUB_ACTION_PATH}/operations/generated_code"
 # BitOps Deployment Config file
 BITOPS_DEPLOY_FILE="${GITHUB_ACTION_PATH}/operations/deployment/bitops.config.yaml"
 # BitOps Code Config File
-BITOPS_CODE_FILE="${GITHUB_ACTION_PATH}/operations/generated_code/bitops.config.yaml"
+##BITOPS_CODE_FILE="${GITHUB_ACTION_PATH}/operations/generated_code/bitops.config.yaml"
 # BitOps Temp file
 BITOPS_CONFIG_TEMP="/tmp/bitops.config.yaml"
 
@@ -111,9 +113,29 @@ BITOPS_CONFIG_TEMP="/tmp/bitops.config.yaml"
 echo -en "
 bitops:
   deployments:
-" > $BITOPS_CONFIG_TEMP
+" > $BITOPS_DEPLOY_FILE
 
-add_terraform_module aws
+# BitOps Config Temp file
+  # Terraform - Generate infra
+  # Will check if there is a tf-state file to work with. If so, will create a destroy step.
+  if [[ $(alpha_only "$TF_STACK_DESTROY") == true ]]; then 
+    if check_statefile aws aws; then
+      add_terraform_module aws
+    fi
+  else
+    if [[ $(alpha_only "$AWS_EC2_INSTANCE_CREATE") == true ]]; then
+      add_terraform_module aws
+    fi
+  fi
+  if [[ $(alpha_only "$AWS_EKS_CREATE") != true ]]; then
+    if check_statefile aws eks; then
+      add_terraform_module eks
+    fi
+  else
+    if [[ $(alpha_only "$AWS_EKS_CREATE") == true ]]; then
+      add_terraform_module eks
+    fi
+  fi
   
   # Ansible Code part
 
@@ -140,9 +162,12 @@ add_terraform_module aws
 
 # Helm part
 
-cp $BITOPS_CONFIG_TEMP $BITOPS_DEPLOY_FILE
-cp $BITOPS_CONFIG_TEMP $BITOPS_CODE_FILE
-rm $BITOPS_CONFIG_TEMP
+if [ ! -s "$BITOPS_CONFIG_TEMP" ]; then
+  rm $BITOPS_DEPLOY_FILE
+ else
+  cat $BITOPS_CONFIG_TEMP >> $BITOPS_DEPLOY_FILE
+  rm $BITOPS_CONFIG_TEMP
+fi
 
 echo "Done with generate_bitops_config.sh"
 exit 0
